refactor: update project name
