+++ back-end/src/main/java/com/bantads/back_end/dto/LoginRequestDTO.java
@@ -1,0 +1,33 @@
+package com.bantads.back_end.dto;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+
+/**
+ * DTO para encapsular os dados da requisição de login.
+ */
+public class LoginRequestDTO {
+
+    @NotBlank(message = "O e-mail é obrigatório")
+    @Email(message = "Formato de e-mail inválido")
+    private String email;
+
+    @NotBlank(message = "A senha é obrigatória")
+    private String password;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+}
