+++ back-end/src/main/java/com/bantads/back_end/dto/LoginResponseDTO.java
@@ -1,0 +1,42 @@
+package com.bantads.back_end.dto;
+
+/**
+ * DTO para a resposta de autenticação.
+ * Envia o token de acesso (JWT) e informações básicas do usuário.
+ */
+public class LoginResponseDTO {
+
+    private String token;
+    private String userName;
+    private String userRole; 
+
+    public LoginResponseDTO(String token, String userName, String userRole) {
+        this.token = token;
+        this.userName = userName;
+        this.userRole = userRole;
+    }
+
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+
+    public String getUserName() {
+        return userName;
+    }
+
+    public void setUserName(String userName) {
+        this.userName = userName;
+    }
+
+    public String getUserRole() {
+        return userRole;
+    }
+
+    public void setUserRole(String userRole) {
+        this.userRole = userRole;
+    }
+}
