+++ back-end/src/main/java/com/bantads/back_end/exception/ResourceNotFoundException.java
@@ -1,0 +1,14 @@
+package com.bantads.back_end.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
+public class ResourceNotFoundException extends RuntimeException {
+	
+	private static final long serialVersionUID = 1L;
+
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
