+++ back-end/src/main/java/com/bantads/back_end/model/Role.java
@@ -1,0 +1,6 @@
+package com.bantads.back_end.model;
+
+public enum Role {
+    ROLE_CLIENT,   
+    ROLE_EMPLOYEE   
+}
