+++ back-end/src/main/java/com/bantads/back_end/security/UserDetailsServiceImpl.java
@@ -1,0 +1,27 @@
+package com.bantads.back_end.security;
+
+import com.bantads.back_end.model.User;
+import com.bantads.back_end.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserDetailsServiceImpl implements UserDetailsService {
+
+    @Autowired
+    private UserRepository userRepository;
+
+    /**
+     * Método que o Spring Security usa para carregar o usuário pelo e-mail (username).
+     */
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new UsernameNotFoundException("Usuário não encontrado com o e-mail: " + email));
+
+        return new UserPrincipal(user);
+    }
+}
