--- back-end/src/main/java/com/remont/back_end/BackEndApplication.java
@@ -1,13 +1,0 @@
-package com.remont.back_end;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BackEndApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BackEndApplication.class, args);
-	}
-
-}
