--- back-end/src/main/java/com/remont/back_end/model/Role.java
@@ -1,6 +1,0 @@
-package com.remont.back_end.model;
-
-public enum Role {
-    ROLE_CLIENT,   
-    ROLE_EMPLOYEE   
-}
