--- back-end/src/main/java/com/remont/back_end/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.remont.back_end.repository;
-
-import com.remont.back_end.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByEmail(String email);
-    
-    boolean existsByEmail(String email);
-}
