--- back-end/src/main/java/com/remont/back_end/service/ClientService.java
@@ -1,17 +1,0 @@
-package com.remont.back_end.service;
-
-import com.remont.back_end.dto.ClientDTO;
-import java.util.List;
-
-public interface ClientService {
-    
-    ClientDTO createClient(ClientDTO clientDTO);
-
-    ClientDTO getClientById(Long id);
-
-    List<ClientDTO> getAllClients();
-
-    ClientDTO updateClient(Long id, ClientDTO clientDTO);
-
-    void deleteClient(Long id);
-}
