--- back-end/src/main/java/com/remont/back_end/service/LoginServiceImpl.java
@@ -1,52 +1,0 @@
-package com.remont.back_end.service;
-
-import com.remont.back_end.dto.LoginResponseDTO;
-import com.remont.back_end.dto.LoginRequestDTO;
-import com.remont.back_end.model.User;
-import com.remont.back_end.repository.UserRepository;
-import com.remont.back_end.security.TokenService;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class LoginServiceImpl implements LoginService 
- {
-    @Autowired
-    private AuthenticationManager authenticationManager;
-
-    @Autowired
-    private TokenService tokenService;
-
-    @Autowired
-    private UserRepository userRepository;
-
-    /**
-     * Processamento de login.
-     */
-    @Override
-    public LoginResponseDTO login(LoginRequestDTO loginRequest) throws AuthenticationException {
-        
-        UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
-                loginRequest.getEmail(),
-                loginRequest.getPassword()
-        );
-
-        Authentication authentication = authenticationManager.authenticate(authToken);
-
-        SecurityContextHolder.getContext().setAuthentication(authentication);
-
-        User user = userRepository.findByEmail(loginRequest.getEmail())
-                .orElseThrow(() -> new UsernameNotFoundException("Usuário não encontrado após autenticação"));
-
-        String token = tokenService.generateToken(user);
-
-        return new LoginResponseDTO(token, user.getName(), user.getRole().name());
-    }
-}
