+++ back-end/src/test/java/com/bantads/back_end/BackEndApplicationTests.java
@@ -1,0 +1,13 @@
+package com.bantads.back_end;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BackEndApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
