--- back-end/src/test/java/com/remont/back_end/BackEndApplicationTests.java
@@ -1,13 +1,0 @@
-package com.remont.back_end;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BackEndApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
